Added file manager functionality to GTK application.
This commit adds the functionality for creating, reading, writing, and deleting files to the GTK application. The main.cpp file has been updated to include functions for each of these actions, as well as buttons in the graphical user interface to trigger them. The application now uses the GTK file chooser dialog to allow the user to select files for reading, writing, and deleting. Error messages are displayed if the file cannot be opened or deleted. Overall, this commit adds basic file management functionality to the application.